refacto: moved sarpostprocessor to sar
